requirements.txt
# Python packages needed for Maia's Mexican Art Machine

# RunPod SDK for connecting to AI art brain
runpod>=1.7.0

# Image processing library
Pillow>=10.0.0

# Load environment variables from .env file (keeps secrets safe!)
python-dotenv>=1.0.0

# For the web version (Vercel serverless functions)
fastapi>=0.94.0
uvicorn>=0.12.0
